refactor: split provider and speech abstractions
